Revert "Swithc to 32 bit"
This reverts commit c89f3c58d9abc8cf6fadf65ecf3f45118d6c4379.

+++ src/store/activity.rs
@@ -7,13 +7,13 @@ use super::JsonStorage;
 pub struct Activity {
     pub name: String,
     pub activity_type: String,
-    pub distance: f32,
-    pub moving_time: u32,
-    pub elapsed_time: u32,
-    pub total_elevation_gain: f32,
+    pub distance: f64,
+    pub moving_time: u64,
+    pub elapsed_time: u64,
+    pub total_elevation_gain: f64,
     pub sport_type: String,
-    pub average_heartrate: Option<f32>,
-    pub max_heartrate: Option<f32>,
+    pub average_heartrate: Option<f64>,
+    pub max_heartrate: Option<f64>,
     pub start_date: DateTime<Utc>,
 }
